requirements.txt
asgiref==3.2.7
attrs==19.3.0
boto3==1.12.36
botocore==1.15.36
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
Django==3.0.4
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-extensions==2.2.9
django-filter==2.2.0
django-js-asset==1.2.2
django-rest-passwordreset==1.1.0
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
idna==2.9
importlib-metadata==1.6.0
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
Markdown==3.2.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
